requirements.txt
alembic==1.7.6
click==8.0.4
DateTime==4.4
Flask==1.1.2
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
PyJWT==2.0.0
python-dotenv==0.19.2
pytz==2021.3
SQLAlchemy==1.4.31
uuid==1.30
Werkzeug==2.0.3
zope.interface==5.4.0
